requirements.txt
PyAutoIt==0.6.0
selenium==3.141.0
urllib3==1.25.9